clarity review - round 1
